Try consolidating some of the copy and paste behaviour
